UniFX v1.0 LTS
v1.0 LTS (passed this exam and am not likely to return to the app)

- needs javafx

The university project for improving OOP skills. 
At the time of writing the program, I did not fully understand JavaFX, so Controller.java was not written very well. Sorry. :)
The GUI isn't very fancy, but it isn't the point of the app (i swear :)).
I created it by myself, without any references. I watched some how-to-do's about JavaFX and that's it. So, this is probably not the most optimal and correct solution, but I wanted to try to make this little system on my own.

this app can:
- view lists of professors, students, subjects, academic plans;
- add new professors, students, subjects (i forget about academic plans, but it's okay, i swear :));
- delete professors, students.